refactor: revert `errorCatcher` logic. Let's keep it one for all mutations.
I'll just extend ValidationError by one more field in next commit. And code will be simplified quite well.
