+++ src/resolvers/updateMany.ts
@@ -16,7 +16,7 @@ import {
 import { toMongoDottedObject } from '../utils/toMongoDottedObject';
 import type { ExtendedResolveParams, GenResolverOpts } from './index';
 import { beforeQueryHelper } from './helpers/beforeQueryHelper';
-import { addErrorCatcherField } from './helpers/addErrorCatcherField';
+import { addErrorCatcherField } from './helpers/errorCatcher';
 
 export default function updateMany<TSource = Document, TContext = any>(
   model: Model<any>,
